Bug 626273 - Correct boundary seek logic.  Change media element IDL floats to doubles per the spec.  Invalidate updated frame after seeking.  Optimize small forward seeks in WebM.  r=cpearce a=roc
UltraBlame original commit: 86c446a17a798ff12c3d9cd8af0064dc60c34959

+++ content/html/content/public/nsHTMLMediaElement.h
@@ -578,7 +578,7 @@ protected:
   LoadAlgorithmState mLoadWaitStatus;
 
   
-  float mVolume;
+  double mVolume;
 
   
   PRUint32 mChannels;
@@ -608,7 +608,7 @@ protected:
 
   
   
-  float mLastCurrentTime;
+  double mLastCurrentTime;
 
   nsRefPtr<gfxASurface> mPrintSurface;
